\name{Seurat_g-class}
\alias{Seurat_g}
\alias{Seurat_g-class}

\docType{class}
\title{GDS-based Seurat Class}
\description{
    The Seurat_g class inherits directly from "Seurat".
}

\section{Slots}{
\code{setClass("Seurat_g", contains="Seurat")}
}

\author{Xiuwen Zheng}
\seealso{
    \code{\link{SCArrayAssay-class}}
}

\keyword{GDS}
\keyword{SingleCell}
